Completed all tests and even played around with the code :)
Want some feedback on the way I did the fight_store function, the way it is right now is my best way of doing it and just wanted to know if it can be improved in any way. Also the code that I commented out is a solution where I use iterators instead. How good is this solution and how does it compare to the way you would make a "one line solution" to this? Or am I just being unnecessary and iterators are just not needed for this? Thanks a lot Forest!